requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.1
decorator==5.1.1
diff-match-patch==20230430
Django
django-allauth==64.0.0
django-celery==3.1.17
django-crispy-forms==2.3
django-filter==24.3
django-geojson==4.1.0
django-guardian==2.4.0
django-humanize==0.1.2
django-import-export==4.1.1
django-leaflet==0.30.1
django-mathfilters==1.0.0
django-role-permissions==3.2.0
django-simple-history==3.7.0
django-slick-reporting==1.3.1
django-tinymce==4.1.0
django-unicorn==0.61.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
Faker==26.3.0
humanize==4.10.0
importlib_metadata==8.2.0
kombu==5.4.0
Markdown==3.6
numpy==2.0.2
openpyxl==3.1.5
orjson==3.10.7
pandas==2.2.2
pillow==10.4.0
prompt_toolkit==3.0.47

python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
shortuuid==1.0.13
simplejson==3.19.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zipp==3.19.2
psycopg2-binary
gunicorn
